requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.10.0
pydantic-settings==2.3.4
sqlalchemy==2.0.44
alembic==1.12.1
psycopg[binary]==3.2.3
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
python-dotenv==1.0.0
httpx==0.27.2
bcrypt==4.1.1
email-validator==2.1.0